1. limit the maximum capturing group to 1,073,741,823
These types are kept as Py_ssize_t:
  - PatternObject.groups
  - MatchObject.lastindex
  - MatchObject.groups

- On 32 bit platform: 36 bytes, no change.  (msvc2022)
- On 64 bit platform: 72 bytes -> 64 bytes. (msvc2022/gcc9.4)

+++ Modules/_sre/sre_lib.h
@@ -512,8 +512,8 @@ typedef struct {
     const SRE_CHAR* ptr;
     const SRE_CODE* pattern;
     Py_ssize_t count;
-    Py_ssize_t lastmark;
-    Py_ssize_t lastindex;
+    int lastmark;
+    int lastindex;
     union {
         SRE_CODE chr;
         SRE_REPEAT* rep;
